refactor: data 계층 패키지 구조 변경
