--- app/src/main/java/woowacourse/shopping/data/cart/CartProductMapper.kt
@@ -1,9 +1,0 @@
-package woowacourse.shopping.data.cart
-
-import woowacourse.shopping.domain.model.CartProduct
-
-fun CartProduct.toEntity() =
-    CartProductEntity(
-        productId = this.product.id,
-        quantity = this.quantity,
-    )
