+++ app/src/main/java/woowacourse/shopping/data/entity/CartProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.cart
+package woowacourse.shopping.data.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
