+++ app/src/main/java/woowacourse/shopping/data/model/PagedResult.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.model
 
 data class PagedResult<T>(
     val items: List<T>,
