+++ app/src/main/java/woowacourse/shopping/data/util/DataMapper.kt
@@ -1,9 +1,7 @@
-package woowacourse.shopping.data.product
+package woowacourse.shopping.data.util
 
 import java.time.Instant
 import java.time.LocalDateTime
 import java.time.ZoneId
 
 fun Long.toLocalDateTime(): LocalDateTime = Instant.ofEpochMilli(this).atZone(ZoneId.systemDefault()).toLocalDateTime()
-
-fun LocalDateTime.toEpochMillis(): Long = this.atZone(ZoneId.systemDefault()).toInstant().toEpochMilli()
