+++ app/src/main/java/woowacourse/shopping/domain/model/RecentProduct.kt
@@ -1,8 +1,16 @@
 package woowacourse.shopping.domain.model
 
+import woowacourse.shopping.data.entity.RecentProductEntity
 import java.time.LocalDateTime
+import java.time.ZoneId
 
 data class RecentProduct(
     val product: Product,
     val viewedAt: LocalDateTime = LocalDateTime.now(),
 )
+
+fun RecentProduct.toEntity() =
+    RecentProductEntity(
+        productId = this.product.id,
+        viewedAt = viewedAt.atZone(ZoneId.systemDefault()).toInstant().toEpochMilli(),
+    )
