+++ app/src/main/java/woowacourse/shopping/domain/repository/ProductRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.domain.repository
 
-import woowacourse.shopping.data.PagedResult
+import woowacourse.shopping.data.model.PagedResult
 import woowacourse.shopping.domain.model.Product
 
 interface ProductRepository {
